+++ src/oop_ext/foundation/callback/_priority_callback.py
@@ -8,6 +8,7 @@ from oop_ext.foundation.decorators import Override
 from oop_ext.foundation.odict import odict
 
 from ._callback import Callback
+from ._callback import _UnregisterContext
 
 
 class PriorityCallback(Callback):
@@ -36,7 +37,7 @@ class PriorityCallback(Callback):
         func: Callable,
         extra_args: Tuple[object, ...] = Callback._EXTRA_ARGS_CONSTANT,
         priority: int = 5,
-    ) -> None:
+    ) -> _UnregisterContext:
         """
         Register a function in the callback.
         :param object func:
@@ -67,3 +68,4 @@ class PriorityCallback(Callback):
             i += 1
 
         callbacks.insert(i, key, (new_info, extra_args))
+        return _UnregisterContext(self, key)